requirements.txt
Django~=2.0.13
djangorestframework==3.10.0
requests==2.22.0
pytest-django==3.5.1
  pytest==5.0.1
django-cors-headers